added existence tests

+++ dist_clean.sh
@@ -2,7 +2,15 @@
 # This script cleans all files, removes the config header, and replaces the
 # Makefile in the project directory with a template
 
-make clean
-rm includes/config.hpp
-rm Makefile
+if [ -f Makefile ]
+then
+    make clean
+    rm Makefile
+fi
+
+if [ -f includes/config.hpp ]
+then
+    rm includes/config.hpp
+fi
+
 cp scripts/Makefile.template Makefile
